requirements.txt
django==5.1
django-filter==24.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
pillow==10.4.0
djoser==2.2.3
django-jalali==7.1.0
kavenegar==1.1.2
